Change display float to double

+++ src/Logs.cpp
@@ -36,7 +36,7 @@ void tree()
     }
 }
 
-void display(float value)
+void display(double value)
 {
     spdlog::info("value: {0}", value);
 }
